Put HadoopTableReader in its own file.
